0.029 taking new approaches to threading in kotlin, corrected a bug, new global thread still doesn't work

+++ app/src/main/java/com/orzechowski/prototyp/database/tutorial/Tutorial.kt
@@ -8,6 +8,6 @@ class Tutorial(
     @PrimaryKey val tutorialId: Long,
                 val tutorialName: String,
                 val authorId: Long,
-                val globalSoundStart: Int,
+                val globalSoundStart: Long,
                 val globalSoundLoop: Boolean,
-                val globalSoundLoopInterval: Int)
+                val globalSoundLoopInterval: Long)
